+++ common/retry/src/main/java/org/occurrent/retry/internal/AfterRetryInfoImpl.java
@@ -22,7 +22,6 @@ import org.occurrent.retry.AfterRetryInfo;
 import org.occurrent.retry.RetryInfo;
 
 import java.time.Duration;
-import java.util.Optional;
 
 record AfterRetryInfoImpl(RetryInfo retryInfo, ResultOfRetryAttempt result, @Nullable Duration backoffBeforeNextRetryAttempt) implements AfterRetryInfo {
     @Override
@@ -66,11 +65,11 @@ record AfterRetryInfoImpl(RetryInfo retryInfo, ResultOfRetryAttempt result, @Nul
     }
 
     @Override
-    public Optional<Duration> getBackoffBeforeNextRetryAttempt() {
+    public Duration getBackoffBeforeNextRetryAttempt() {
         if (wasSuccessfulRetryAttempt()) {
-            return Optional.empty();
+            return null;
         } else {
-            return Optional.ofNullable(backoffBeforeNextRetryAttempt);
+            return backoffBeforeNextRetryAttempt;
         }
     }